ver_modelos.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

print("--- BUSCANDO MODELOS DISPONÍVEIS PARA VOCÊ ---")

try:
    for m in genai.list_models():
        if 'generateContent' in m.supported_generation_methods:
            print(f"Nome valido: {m.name}")

except Exception as e:
    print(f"Erro ao listar: {m.name}")