requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
google-genai
python-dotenv
pydantic
httpx